refactor StrategyPerformanceChart part 1:
extract ChartContainer outer/parent component
