+++ src/lib/chart/index.ts
@@ -1,6 +1,7 @@
 export * from './quote-feed';
 export * from './time-bucket-converters';
 export { default as ChartActivityTracker } from './ChartActivityTracker';
+export { default as ChartContainer } from './ChartContainer.svelte';
 export { default as ChartIQ } from './ChartIQ.svelte';
 export { default as ChartLinker } from './ChartLinker';
 export { default as HudRow } from './HudRow.svelte';
@@ -8,4 +9,3 @@ export { default as HudMetric } from './HudMetric.svelte';
 export { default as Marker } from './Marker.svelte';
 export { default as PairCandleChart } from './PairCandleChart.svelte';
 export { default as ReserveInterestChart } from './ReserveInterestChart.svelte';
-export { default as StrategyPerformanceChart } from './StrategyPerformanceChart.svelte';
