+++ src/routes/strategies/[strategy]/(nav)/netflow/+page.svelte
@@ -3,7 +3,7 @@
 -->
 <script lang="ts">
 	import WebChart from '../../WebChart.svelte';
-	import { StrategyPerformanceChart } from '$lib/chart';
+	import { ChartContainer } from '$lib/chart';
 	import { formatDaysAgo, formatDollar } from '$lib/helpers/formatters';
 
 	export let data;
@@ -12,11 +12,11 @@
 
 <section class="tvl">
 	<p>Displaying live trading metrics. This strategy has been live <strong>{formatDaysAgo(startedAt)}</strong>.</p>
-	<StrategyPerformanceChart title="Total value locked" data={tvlChart.data} formatValue={formatDollar}>
+	<ChartContainer title="Total value locked" data={tvlChart.data} formatValue={formatDollar}>
 		Learn more about
 		<a class="body-link" href={tvlChart.help_link}>Total value locked</a>
 		metric and how it is calculated.
-	</StrategyPerformanceChart>
+	</ChartContainer>
 	<WebChart name="Netflow" webChart={netflowChart} charType="bar" />
 </section>
